# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ruckus Conference Program Operations API
  version: 1.0.0
  summary: Conference program operations - CFP, review, speakers, schedule, comms, CRM and public program feeds.
  description: >-
    HTTP API for Ruckus, an open-source conference program-operations tool. Paths are
    generated from the live Hono route registrations in src/app.ts and the src route
    modules, so path and method coverage is exact. Request and response bodies are
    intentionally loose object schemas with descriptions: handlers accept and return
    JSON envelopes shaped as a data property, and the demo does not enforce strict
    field level typing. Identity is demo persona simulation, not authentication.
  license:
    name: MIT
servers:
  - url: https://ruckus.to
    description: Live demo deployment
  - url: http://localhost:8787
    description: Local development API
...
security: []
paths:
...
  "/api/content/files/{fileId}/comments":
    parameters:
      - name: fileId
        in: path
        required: true
        schema:
          type: string
        description: "fileId path parameter."
    post:
      tags: [speakers]
      operationId: post_api_content_files_fileId_comments
      summary: "Registered in contentRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '200':
          description: Successful response for operations that update in place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  "/api/content/files/{fileId}/versions/{versionId}":
    parameters:
      - name: fileId
        in: path
        required: true
        schema:
          type: string
        description: "fileId path parameter."
      - name: versionId
        in: path
        required: true
        schema:
          type: string
        description: "versionId path parameter."
    get:
      tags: [speakers]
      operationId: get_api_content_files_fileId_versions_versionId
      summary: "Registered in contentRoutes.ts. See docs/API.md for role and behaviour notes."
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  "/api/events/{eventId}/content":
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
        description: "eventId path parameter."
    get:
      tags: [speakers]
      operationId: get_api_events_eventId_content
      summary: "Registered in contentRoutes.ts. See docs/API.md for role and behaviour notes."
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  "/api/events/{eventId}/content/export":
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
        description: "eventId path parameter."
    get:
      tags: [speakers]
      operationId: get_api_events_eventId_content_export
      summary: "Registered in contentRoutes.ts. See docs/API.md for role and behaviour notes."
      responses:
        '200':
          description: Successful response.
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags: [speakers]
      operationId: post_api_events_eventId_content_export
      summary: "Registered in contentRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '201':
          description: Successful response.
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '200':
          description: Successful response for operations that update in place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  "/api/events/{eventId}/content/files/{fileId}/approval":
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
        description: "eventId path parameter."
      - name: fileId
        in: path
        required: true
        schema:
          type: string
        description: "fileId path parameter."
    patch:
      tags: [speakers]
      operationId: patch_api_events_eventId_content_files_fileId_approval
      summary: "Registered in contentRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  "/api/events/{eventId}/content/files/{fileId}/versions/{versionId}":
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
        description: "eventId path parameter."
      - name: fileId
        in: path
        required: true
        schema:
          type: string
        description: "fileId path parameter."
      - name: versionId
        in: path
        required: true
        schema:
          type: string
        description: "versionId path parameter."
    get:
      tags: [speakers]
      operationId: get_api_events_eventId_content_files_fileId_versions_versionId
      summary: "Registered in contentRoutes.ts. See docs/API.md for role and behaviour notes."
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  "/api/events/{eventId}/content/history/{historyId}/restore":
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
        description: "eventId path parameter."
      - name: historyId
        in: path
        required: true
        schema:
          type: string
        description: "historyId path parameter."
    post:
      tags: [speakers]
      operationId: post_api_events_eventId_content_history_historyId_restore
      summary: "Registered in contentRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '200':
          description: Successful response for operations that update in place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  "/api/events/{eventId}/content/reminders":
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
        description: "eventId path parameter."
    post:
      tags: [speakers]
      operationId: post_api_events_eventId_content_reminders
      summary: "Registered in contentRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '201':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '200':
          description: Successful response for operations that update in place.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  "/api/events/{eventId}/content/sessions/{sessionId}":
    parameters:
      - name: eventId
        in: path
        required: true
        schema:
          type: string
        description: "eventId path parameter."
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
        description: "sessionId path parameter."
    patch:
      tags: [speakers]
      operationId: patch_api_events_eventId_content_sessions_sessionId
      summary: "Registered in contentRoutes.ts. See docs/API.md for role and behaviour notes."
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JsonObject'
      responses:
        '200':
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataEnvelope'
        '400':
          description: Validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Demo role or ownership scope rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Event, record or route target was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  schemas:
    JsonObject:
      type: object
      additionalProperties: true
      description: >-
        Loosely typed JSON request body. Handlers read the fields they need and ignore
        the rest; see docs/API.md for the fields each operation consumes.
    DataEnvelope:
      type: object
      additionalProperties: true
      description: >-
        Standard success envelope. Most JSON endpoints return a data property holding an
        object or array; some also return meta describing counts or provider outcomes.
      properties:
        data:
          description: Operation payload, either an object or an array.
        meta:
          type: object
          additionalProperties: true
          description: Optional counts, filters or provider status details.
    ErrorEnvelope:
      type: object
      additionalProperties: true
      description: Standard failure envelope returned by the shared error helper.
      properties:
        error:
          type: object
          additionalProperties: true
          properties:
            code:
              type: string
              description: Short machine readable code such as NOT_FOUND or VALIDATION_ERROR.
            message:
              type: string
              description: Human readable explanation.